# Conflicts:
#	NewLaserProject/Classes/Process/GeneralLaserProcess.cs

+++ NewLaserProject/Data/Models/MaterialEntRule.cs
@@ -4,8 +4,14 @@ namespace NewLaserProject.Data.Models
 {
     public class MaterialEntRule:BaseEntity
     {
-        public double Offset { get; set; }
-        public double Width { get; set; }
+        /// <summary>
+        /// Measured in um
+        /// </summary>
+        public int Offset { get; set; }
+        /// <summary>
+        /// Measured in um
+        /// </summary>
+        public int Width { get; set; }
         [Browsable(false)]
         public int MaterialId { get; set; }
         [Browsable(false)]
